Implement higher-precision calculations.

+++ cleanup/filter/erode.go
@@ -1,6 +1,6 @@
 package filter
 
-func max(a, b, c byte) byte {
+func max(a, b, c float32) float32 {
 	if a >= b {
 		if a >= c {
 			return a
